Funcionalidad agregar, editar y eliminar dentro modal anotadores

+++ Vacaciones/Controllers/EjecutivosYPlantaController.cs
@@ -126,7 +126,7 @@ namespace Vacaciones.Controllers
                     apllds_slctnte = ApellidosEmpleado,
                     fcha_inco_vccns = Convert.ToDateTime(FechaInicio),
                     fcha_fn_vcc = Convert.ToDateTime(FechaFin),
-                    nmro_ds = int.Parse(NumeroDias),
+                    nmro_ds = double.Parse(NumeroDias),
                     sbdo_hbl = oRespuestaSAPModels.Details[0].SabadoHabil == "NO" ? false : true,
                     fcha_hra_aprvc = DateTime.Now,
                     fcha_hra_rgstro_nvdd = DateTime.Now,
